Cleaned up pulseblaster files.
